go core

+++ main.go
@@ -3,16 +3,27 @@ package main
 import (
 	"fmt"
 	"head/main_"
+	"head/main_/func_all"
 	"net/http"
+	"os/exec"
 	"strconv"
 )
 
 func main() {
-	port := main_.FindFreePort()
+	config, err := func_all.LoadConfig_start(main_.Main_config)
+	if err != nil {
+		fmt.Printf("Не вдалося завантажити конфігурацію: %v\n", err)
+		return
+	}
+
+	port := func_all.FindFreePort()
 	portStr := ":" + strconv.Itoa(port)
 
-	main_.Write_config_core(portStr)
-	cmd := main_.StartShellWeb()
+	var cmd *exec.Cmd
+	if config.Visualization == 1 {
+		func_all.Write_config_core(portStr)
+		cmd = func_all.StartShellWeb()
+	}
 
 	http.Handle("/static/", http.StripPrefix("/static/", http.FileServer(http.Dir("static"))))
 
@@ -31,12 +42,14 @@ func main() {
 	http.HandleFunc("/visualization", main_.Post_config_change)
 
 	fmt.Printf("Сервер працює на порту %d\n", port)
-	err := http.ListenAndServe(portStr, nil)
+	err = http.ListenAndServe(portStr, nil)
 	if err != nil {
 		fmt.Printf("Помилка запуску сервера: %v\n", err)
 	}
 
-	if err := cmd.Process.Kill(); err != nil {
-		fmt.Printf("Не вдалося завершити shell_web.exe: %v\n", err)
+	if cmd != nil {
+		if err := cmd.Process.Kill(); err != nil {
+			fmt.Printf("Не вдалося завершити shell_web.exe: %v\n", err)
+		}
 	}
 }
